Switch from long to int
git-svn-id: https://svn-server/svn/maayanlab/trunk/Enrichr@1731 998562b1-c221-4fe5-b3d3-2fd2107a5f19

+++ src/main/java/edu/mssm/pharm/maayanlab/Enrichr/Shortener.java
@@ -5,7 +5,7 @@ public class Shortener {
 	private static final String alphabet = "123456789abcdefghijkmnopqrstuvwxyzABCDEFGHJKLMNPQRSTUVWXYZ";
 	private static final int base = 58;
 	
-	public static String encode(long number) {
+	public static String encode(int number) {
 		StringBuilder linkId = new StringBuilder();
 		
 		do {
@@ -18,12 +18,12 @@ public class Shortener {
 		return linkId.toString();
 	}
 	
-	public static long decode(String linkId) {
-		long number = 0;
+	public static int decode(String linkId) {
+		int number = 0;
 		
 		for (int i = 0; i < linkId.length(); i++) {
 			int position = alphabet.indexOf(linkId.charAt(i));
-			number += position * (long) Math.pow(base, linkId.length() - i - 1);
+			number += position * (int) Math.pow(base, linkId.length() - i - 1);
 		}
 		
 		System.out.println("");
